requirements.txt
# Core dependencies - install these first
setuptools>=65.0.0
wheel>=0.37.0
pip>=23.0.0

# Basic data processing
numpy>=1.21.0
pandas>=1.5.0

# PDF processing
PyMuPDF>=1.20.0
pdfplumber>=0.7.0

# Machine learning and embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.2.0

# Web framework
streamlit>=1.28.0

# Deep learning (optional - will install with sentence-transformers)
torch>=1.9.0
transformers>=4.20.0
